added workaround for Horizons SBDB API inconsistency, added orbit quality check
The Horizons small body database doesn't always use a consistent data structure with regards to the best-fit orbit and the covariance matrix. The code now tries a second way to grab the best-fit orbit from the SBDB query results if it doesn't find the best-fit within the covariance matrix data as expected (and returns an error if the other place the best-fit can be found isn't for the same epoch as the covariance matrix). There is also now an observational arc length check to prevent attempts to clone objects with huge uncertainties.