added test for planet query, updated small body query with packed designations, cleaned up style
There's now a tests folder with a test script to check if the horizons query for the planets is working right (can't do a test for the small bodies as the best-fit orbit epoch is unknown and can change). I also added a routine that can convert from an unpacked MPC temporary designation to a packed designation as this should prevent some of the weird Horizons url-query errors that have popped up. The Horizons query now uses the packed designation as needed. I also updated the horizons_api.py file to standard python documentation formatting.