corrected error in hard-coded planetary radii
there was a m/km issue in the planet radii (thanks Sarah Greenstreet for catching it!) and the hard coded values were off by 10^3. The planet radii are now fixed! (This would only have affected collision calculations in any simulations run with the rebound tools, not the close-encounter triggering or anything else). Also double checked and made tiny modifications to the Earth-moon and Mars system GM values to be up to date with the cited reference